pan/va: Don't truncate slots
Causes BARRIER not to work.

Fixes: f45654af595 ("pan/va: Add packing routines")
Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/15756>

+++ src/panfrost/bifrost/valhall/va_pack.c
@@ -617,7 +617,7 @@ va_pack_instr(const bi_instr *I, unsigned flow)
    hex |= ((uint64_t) va_select_fau_page(I)) << 57;
 
    if (info.slot) {
-      unsigned slot = (I->op == BI_OPCODE_BARRIER) ? 7 : 0;
+      uint64_t slot = (I->op == BI_OPCODE_BARRIER) ? 7 : 0;
       hex |= (slot << 30);
    }
